Try delete

+++ application/models/usermodel.php
@@ -181,9 +181,13 @@ class Usermodel extends CI_Model
         return $this->db->affected_rows() > 0;
     }
 
-    public function deleteUser()
+    public function deleteUser($password)
     {
+        if (crypt($password, $user['password']) != $user['password']) {
+            return false;
+        }
         $this->db->delete('users', array('userid' => $this->session->userdata('userid')));
+        return true;
     }
 
     public function like($user)
